+++ src/Animation.c
@@ -21,7 +21,7 @@ animation load_animation(const char** locations, int size, vec2 position)
     };
     
 }
-void play_animation(animation* data, int target_fps, float scale, float rotation, tint color_drawn)
+void play_animation(animation* data, long target_fps, float scale, float rotation, tint color_drawn)
 {
     for (data->current_frame <= data->frames; data->current_frame++;)
     {
